+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.0.17]
+
 ### Added
 
 - Configuration option to show/hide the Settings menu item in the Drawer
@@ -220,7 +222,8 @@ Initial Version
     - displaying the question number in the details screen
     - displaying the question title in the details screen
 
-[unreleased]: https://github.com/fliturgy/catechism/compare/v0.0.16...HEAD
+[unreleased]: https://github.com/fliturgy/catechism/compare/v0.0.17...HEAD
+[0.0.17]: https://github.com/fliturgy/catechism/compare/0.0.16...v0.0.17
 [0.0.16]: https://github.com/fliturgy/catechism/compare/0.0.15...v0.0.16
 [0.0.15]: https://github.com/fliturgy/catechism/compare/0.0.14...v0.0.15
 [0.0.14]: https://github.com/fliturgy/catechism/compare/v0.0.13...0.0.14